asos-backend/src/main/java/es/andim/asos/domain/Association.java
package es.andim.asos.domain;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

import lombok.AllArgsConstructor;

@Component
@AllArgsConstructor
public class Association {

    @Autowired
    private MemberRepository memberRepository;

    public List<Member> getAllMembers() {
        return memberRepository.findAll();
    }

    public Member addNewMember(Member member) throws MemberAlreadyExistsException {
        return memberRepository.addMember(member);
    }
}
